pyproject.toml
[project]
name = "lobsterlogreporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.0",
    "pyqtdarktheme>=2.1.0",
    "pyside6-essentials>=6.10.2",
]
